+++ plugins/image-trust/pkg/discovery/result.go
@@ -2,11 +2,9 @@ package discovery
 
 import (
 	"github.com/fairwindsops/insights-plugins/plugins/image-trust/pkg/models"
-	"github.com/fairwindsops/insights-plugins/plugins/image-trust/pkg/registry"
 )
 
-// Result contains images discovered in scope and pull secrets referenced by those workloads.
+// Result contains images discovered in scope.
 type Result struct {
-	Images         []models.DiscoveredImage
-	PullSecretRefs []registry.PullSecretRef
+	Images []models.DiscoveredImage
 }