+++ plugins/image-trust/pkg/registry/pullsecret_ref.go
@@ -1,7 +1,0 @@
-package registry
-
-// PullSecretRef identifies a kubernetes.io/dockerconfigjson secret referenced by a workload.
-type PullSecretRef struct {
-	Namespace string
-	Name      string
-}